Kernel.cs
The expression evaluation and output variable value are included. Haven't included the string concatenation yet. Also, when entering data it has to be in the format as "alpha = 23*3", with the only spaces being on bot sides of the equal sign. On command out, the same applies for example, "alpha out" will show correct result but "alpha out ", with space at the end will not. Another issue I have to look into is using different arithmetic operators on an expression for example "23+4*7-10".